test(fetch): merge client-name scope tests into telemetry_scope
Consolidate the http.client.name instrumentation-scope tests into
telemetry_scope.rs alongside the fetch.runtime / fetch.transport scope
tests, reusing the existing scope_attribute_for_metric helper, and remove
the standalone client_name_metric.rs.
