+++ crates/fetch/tests/telemetry_scope.rs
@@ -187,6 +187,50 @@ async fn tokio_transport_connection_metric_scope_attribute() {
     );
 }
 
+#[cfg_attr(miri, ignore)]
+#[tokio::test]
+async fn client_name_set_before_meter_provider_is_reported() {
+    let (exporter, provider) = exporter_and_provider();
+
+    let client = HttpClient::builder_fake(StatusCode::OK, &Clock::new_frozen())
+        .insecure_allow_http()
+        .name("named_before_provider")
+        .meter_provider(provider.clone())
+        .build();
+
+    client.get("http://example.com").fetch().await.unwrap();
+    provider.force_flush().unwrap();
+
+    assert_eq!(
+        scope_attribute_for_metric(&exporter, "http.client.request.duration", "http.client.name"),
+        Some("named_before_provider".to_owned()),
+        "a client name set before meter_provider must be reported as the http.client.name scope attribute"
+    );
+}
+
+#[cfg_attr(miri, ignore)]
+#[tokio::test]
+async fn client_name_set_after_meter_provider_is_reported() {
+    let (exporter, provider) = exporter_and_provider();
+
+    // Setting the meter provider *before* the name exercises the deferred-scope path: the
+    // custom meter is only materialized at build time, so the later name still applies.
+    let client = HttpClient::builder_fake(StatusCode::OK, &Clock::new_frozen())
+        .insecure_allow_http()
+        .meter_provider(provider.clone())
+        .name("named_after_provider")
+        .build();
+
+    client.get("http://example.com").fetch().await.unwrap();
+    provider.force_flush().unwrap();
+
+    assert_eq!(
+        scope_attribute_for_metric(&exporter, "http.client.request.duration", "http.client.name"),
+        Some("named_after_provider".to_owned()),
+        "a client name set after meter_provider must be reported as the http.client.name scope attribute"
+    );
+}
+
 /// Transport handler that returns a canned `200 OK`.
 struct OkHandler {
     body_builder: http_extensions::HttpBodyBuilder,
